clinic-backend/requirements.txt
fastapi
beanie
python-dotenv
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
python-jwt
cloudinary
python-multipart